// typings/globals.d.ts
export type SearchData = {
  [key: string]: any,
  city: string,
  region: string,
  zip: string,
  tuin: boolean,
  garage: boolean,
  startDate: string | null,
  endDate: string | null,
  minPrice: number | null,
  maxPrice: number | null,
  minRooms: number | null,
  maxRooms: number | null,
  minParcel: number | null,
  maxParcel: number | null,
  minLiving: number | null,
  maxLiving: number | null,
  dakterras: boolean | null,
  klushuis: boolean,
  nieuwbouw: boolean,
  minBouwjaar: number | null,
  maxBouwjaar: number | null,
  energyLabel: string | null,
}